requirements.txt
Django==2.1.3
django-cleanup==2.1.0
django-treebeard==4.3
djangorestframework==3.9.0
mysqlclient==1.3.13